+++ Tests/CodexBarTests/CLICardsClaudeSwapTests.swift
@@ -373,14 +373,7 @@ struct CLICardsClaudeSwapTests {
         ]}
         JSON
         """
-        // Executing a file this process just wrote races with every other test
-        // that spawns a child: the fork inherits the still-open write descriptor
-        // and execve fails with ETXTBSY. Write the body as data instead, and
-        // execute the checked-in trampoline that reads it.
-        try Data(script.utf8).write(to: executable.appendingPathExtension("sh"))
-        try FileManager.default.createSymbolicLink(
-            at: executable,
-            withDestinationURL: Self.execTrampoline())
+        try FakeExecutable.install(script, at: executable)
 
         let output = await CLIClaudeSwapCards.fetch(
             eligible: true,
@@ -423,19 +416,5 @@ struct CLICardsClaudeSwapTests {
         #expect(output.cards == ambient.cards)
         #expect(output.cardFailures.last?.accountLabel == "claude-swap")
         #expect(output.exitCode != .success)
-    }
-
-    private static func execTrampoline() throws -> URL {
-        var directory = URL(filePath: #filePath).deletingLastPathComponent()
-        for _ in 0..<12 {
-            let candidate = directory.appending(path: "Tests/CodexBarTests/Fixtures/Scripts/exec-trampoline.sh")
-            if FileManager.default.fileExists(atPath: candidate.path(percentEncoded: false)) {
-                return candidate
-            }
-            directory.deleteLastPathComponent()
-        }
-        throw NSError(domain: "CLICardsClaudeSwapTests", code: 1, userInfo: [
-            NSLocalizedDescriptionKey: "Could not locate exec-trampoline.sh from \(#filePath)",
-        ])
     }
 }
