+++ TestsLinux/CLICardsClaudeSwapTests.swift
@@ -372,8 +372,7 @@ struct CLICardsClaudeSwapTests {
         ]}
         JSON
         """
-        try Data(script.utf8).write(to: executable)
-        try FileManager.default.setAttributes([.posixPermissions: 0o755], ofItemAtPath: executable.path)
+        try FakeExecutable.install(script, at: executable)
 
         let output = await CLIClaudeSwapCards.fetch(
             eligible: true,
